refactor; move cdm
