+++ tests/tests_integration/test_api/test_data_modeling/test_core_model.py
@@ -7,7 +7,6 @@ import pytest
 from _pytest.mark import ParameterSet
 
 from cognite.client import CogniteClient
-from cognite.client.data_classes.cdm import v1 as cdm
 from cognite.client.data_classes.data_modeling import (
     Space,
     SpaceApply,
@@ -16,6 +15,7 @@ from cognite.client.data_classes.data_modeling import (
     TypedNode,
     TypedNodeApply,
 )
+from cognite.client.data_classes.data_modeling.cdm import v1 as cdm
 
 DATA_SPACE = "python_sdk_core_v1_test_space"
